(svn r19333) -Fix (r19328): int versus uint comparison warning fixed.

+++ src/airport_gui.cpp
@@ -261,7 +261,7 @@ public:
 			if (!as->IsAvailable()) {
 				GfxFillRect(r.left + 1, y + 1, r.right - 1, y + this->line_height - 2, 0, FILLRECT_CHECKER);
 			}
-			DrawString(r.left + WD_MATRIX_LEFT, r.right + WD_MATRIX_RIGHT, y + WD_MATRIX_TOP, as->name, (i == _selected_airport_index) ? TC_WHITE : TC_BLACK);
+			DrawString(r.left + WD_MATRIX_LEFT, r.right + WD_MATRIX_RIGHT, y + WD_MATRIX_TOP, as->name, ((int)i == _selected_airport_index) ? TC_WHITE : TC_BLACK);
 			y += this->line_height;
 		}
 	}
